♻️ Renamed cell components for better transperancy
